Expand `~` in input and ouput mappings
Bash expansion of `~` to a users homer directory can be a little finicky. For
example:

```sh
./piper -c task.yml -i my_app=~/workspace/my-app
```

will work, while

```sh
./piper -c task.yml -i my-app=~/workspace/my-app
```

According to [this StackOverflow answer][0], the shell expansion works in the
first example because `my_app=~/workspace/my-app` resembles a variable
assignment. In the second example, `my-app=~/workspace/my-app`  does not
resemble a valid variable assignment.

The problem is a little trickier in shells like `zsh` or `sh` (see [this SO
answer][1] for more details). To be clear, this is not an issue with `piper`
but the semantics of shell expansion. The proposed feature is purely a
convenience when using `piper` interactively (as opposed to inside a shell
script).

[0]: https://stackoverflow.com/a/15504442
[1]: https://unix.stackexchange.com/a/373532