Automatically map resources to dirs in work dir
For Concourse tasks that have several input/output resources, it can be
tedious to write out explicit resource mappings via `-i` or `-o` when
using piper from the command line (versus inside of a bash script). This
commit adds functionality to look for input and output resources in the
current working directory, matching on the resource name.